Crop large images to the region where there are actually cells
This makes it possible to train on large microscopy images, where the organoid only occupies a small part of the image.

Previously, OrganoidTracker would load the entire 3D image, then create a labels image of the same size. Then it would take a crop around the location where there were actually labels. However, the original image was kept in memory, as if you slice in numpy, it keeps a reference to the original image.

Now it loads the entire 3D image, crops it, makes a copy of the pixels of the crop so that the original image can be erased from memory, and then creates the label image specifically only for the area of the crop.