Avoid costly merging and splitting up of tracks
When loading from the new file format, we already know that our track structure is correct, so no point to first merge two tracks and then later split them up again.

For a file with a ton of short tracks (output of predict_links), the loading time decreased from 32 minutes to just 57 seconds.