Fixes three bugs in the file saving for the new format
The first one is a small mistake that shows up when min_time_point is not zero. The two other are more serious and give strange shifts in linkdata when not all links are assigned the same metadata.
The fix is somewhat ugly admittedly, would maybe be better to treat missing data more consistenly.