Move all error display settings to warning limits
Now all filtering is saved to the Experiment object, and done in the cell_error_finder module.

All error filtering settings can now be found in the Errors menu of the link and position editor.

This should make things more consistent, and avoid the issue that different visualizers have different error filtering.