Fix adding a mask that doesn't intersect with the selected position
Every mask is only assigned to a position if that position falls within the mask. When you started drawing a mask on a different XY plane, the position would be outside. The editor now adds a single pixel at the position, so that the position will always be assigned to the mask.