Allow running all tracking steps at once on the Keras3 branch
@maxbetjes had implemented this for the master branch, but since the tracking scripts work a bit differently on the Keras3 branch (with autlist files), I had to re-implement it.

You can now load in 10 datasets, and run all tracking steps in sequence on all of them. First it would detect the positions in all datasets, then the divisions, etc.

Also updated the tracking scripts to skip already processed files. So if the tracking is interrupted, you can just restart the script, and it will pick up where it left off. Only for the last tracking step, organoid_tracker_create_tracks, this hasn't been implemented yet.

And fixed some bugs with output folders being created relative to the folder of the autlist file, instead of relative to the folder you ran the script from.