Replace CZI file loader by a more modern one
The old one had the advantage of being just one Python file, but it was getting quite outdated. (Last updated in 2019.) Nowadays, I would be surprised if a fresh CZI file did load correctly. So it was time to switch to another library. The choice fell on aicspylibczi, since it's available on Conda-Forge.