Get rid of experiment.link_data too
It now only exists as a shim for backwards compatibility. link_data was less involved to replace than position_data, since it is used less.

Also updated all the tests to account for that you now need to add the relevant positions to the collection before you can set metadata.