Add drag and drop support for images and tracking data
Over the years, I've seen so many people assume this works, that I thought it would be a good idea to actually make it work.

The main change is that OrganoidTracker is now aware of which file formats it supports. Plugins can now register support for file types, which end up in `window.registry`. Once a file is dragged in, OrganoidTracker checks if it has a FileLoader that can load it, and if yes, that loader gets executed.

The existing loading dialogs have been refactored to use this new FileLoader system.

All default file loaders (LIF, TIF, ND2, tracking files, etc.) are registered by `plugin_loaders_for_default_file_formats.py`.